requirements.txt
Flask
Flask-MySQLdb==1.0.1
Flask-Session==0.5.0
Flask-WTF==1.2.1
bcrypt==4.1.2
PyMySQL==1.1.0
mysql-connector-python==8.3.0
reportlab==4.1.0
qrcode==7.4.2
Pillow==10.2.0
python-barcode==0.15.1
requests==2.31.0
python-dotenv==1.0.0
cryptography==42.0.2
WTForms==3.1.2
Werkzeug==3.0.1
itsdangerous==2.1.2
gunicorn
